requirements.txt
discord.py
BeautifulSoup4
asyncio
selenium